requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
appdirs==1.4.4
async-timeout==5.0.1
attrs==24.3.0
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
Deprecated==1.2.15
frozenlist==1.5.0
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
ipinfo==5.1.1
lxml==5.3.0
MouseInfo==0.1.3
multidict==6.1.0
numpy==2.2.1
opencv-python==4.10.0.84
packaging==24.2
pillow==11.0.0
propcache==0.2.1
PyAutoGUI==0.9.54
pycparser==2.22
PyGetWindow==0.0.9
PyGithub==2.5.0
PyJWT==2.10.1
PyMsgBox==1.0.9
PyNaCl==1.5.0
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pyshark==0.6
pytweening==1.2.0
requests==2.32.3
scapy==2.6.1
smmap==5.0.2
SoundCard==0.4.3
soundfile==0.12.1
termcolor==2.5.0
typing_extensions==4.12.2
urllib3==2.3.0
wrapt==1.17.0
yarl==1.18.3
